chore: move RecursiveCheckEquals to AnimatorOptimizerTestBase
